backend/requirements.txt
fastapi
uvicorn[standard]
python-jose
passlib[bcrypt]
pydantic
PyJWT